Use pkgsrc-wip weston and bluez-alsa
Xfce partially built, but x11 in pkgsrc has no Linux modeset or input.
Weston runs with a subset of the deps I'd have to maintain for those, so
package and send it to pkgsrc-wip instead. Along with a new bluez which is
okay with musl and which my ctl script can befriend with my earphones,
ffmpeg4 and alsa-utils. Build liberation-ttf, nice system font family.